chore(web): delete legacy JobProgressCard + dead CSS (sc-2093)
Final slice of epic sc-2080. All consumers migrated to
WorkerProgressCard in sc-2088..2092, so the legacy components and
their CSS can go:

Deleted:
  - apps/web/src/components/JobProgress.jsx (JobProgressCard +
    useLiveJobElapsedSeconds — the hook moved to WorkerProgressCard
    since it's the only remaining consumer).
  - .local-job-card / .local-job-main / .local-job-actions CSS blocks.
  - .job-row / .job-main / .job-meta / .job-actions / .job-message
    CSS blocks plus the .job-row.<status> status-badge overrides.
  - .training-live-card / .training-live-head / .training-live-meta /
    .training-sample-grid / .training-sample-tile CSS blocks (the
    .training-live-panel / .training-live-title / .training-live-list
    wrappers stay — TrainingStudio still uses them as the panel
    heading + stack container).
  - .job-actions button (and :hover) entry from the shared button
    styling cascade — purely the deleted JobRow's action bar.
  - .training-sample-grid from the mobile media-query rule.

Kept:
  - .local-job-stack — composed alongside .worker-progress-card-stack
    in Image / Video / Document studios for back-compat.
  - .local-job-group — DocumentStudio + ImageStudio still wrap the
    card in an article with this class.
  - .progress-track / .worker-meter / .status-badge / .error-text
    — primitives the new card composes on.

Test: updated one main.test.jsx assertion that pinned the now-deleted
.local-job-card to its new equivalent (.worker-progress-card).

215/215 vitests still green; Vite build clean. CSS bundle shrinks
from 104.60 kB to 102.03 kB (~2.5 kB after legacy purge).

Story: sc-2093
