--- apps/web/src/components/JobProgress.jsx
@@ -1,93 +1,0 @@
-import React, { useEffect, useState } from "react";
-import { terminalStatuses } from "../constants.js";
-import { errorStatuses } from "../jobTypes.js";
-import { formatSeconds, liveElapsedSeconds, percent } from "../formatting.js";
-
-function formatJobType(type) {
-  const label = String(type ?? "job").replaceAll("_", " ");
-  return label.charAt(0).toUpperCase() + label.slice(1);
-}
-
-function jobTitle(job) {
-  return job.payload?.prompt ?? job.payload?.name ?? job.payload?.loraId ?? job.payload?.modelName ?? job.payload?.modelId ?? job.id;
-}
-
-function jobMessage(job) {
-  if (job.error || job.message) {
-    return job.error ?? job.message;
-  }
-  if (job.status === "queued") {
-    return "Queued and waiting for an eligible worker.";
-  }
-  if (job.status === "completed") {
-    return "Finished. Fetching result...";
-  }
-  if (job.stage && job.stage !== job.status) {
-    return `Stage: ${formatJobType(job.stage)}.`;
-  }
-  return "";
-}
-
-export function useLiveJobElapsedSeconds(job) {
-  const active = !terminalStatuses.has(job.status) && Boolean(job.startedAt);
-  const [nowMs, setNowMs] = useState(() => Date.now());
-
-  useEffect(() => {
-    if (!active) {
-      return undefined;
-    }
-    const timer = window.setInterval(() => setNowMs(Date.now()), 1000);
-    return () => window.clearInterval(timer);
-  }, [active, job.startedAt]);
-
-  return liveElapsedSeconds(job, nowMs);
-}
-
-export function JobProgressCard({ job, label, onOpenQueue, onCancel }) {
-  const isError = errorStatuses.has(job.status);
-  const canCancel = !terminalStatuses.has(job.status);
-  const progressLabel = percent(job.progress);
-  const message = jobMessage(job);
-  const elapsedSeconds = useLiveJobElapsedSeconds(job);
-  return (
-    <article className={`local-job-card ${job.status}`}>
-      <div className="local-job-main">
-        <div>
-          <p className="eyebrow">{label ?? formatJobType(job.type)}</p>
-          <h3>{jobTitle(job)}</h3>
-        </div>
-        <span className={`status-badge ${job.status}`}>{job.status}</span>
-      </div>
-      <div className="progress-track" aria-label={`${progressLabel} complete`}>
-        <span style={{ width: progressLabel }} />
-      </div>
-      <div className="job-meta">
-        <span>{formatJobType(job.stage ?? job.status)}</span>
-        <span>{formatSeconds(elapsedSeconds)}</span>
-        <span>GPU {job.assignedGpu ?? job.requestedGpu ?? "auto"}</span>
-      </div>
-      {message ? <p className={isError ? "job-message error-text" : "job-message"}>{message}</p> : null}
-      {onCancel && canCancel ? (
-        <div className="local-job-actions">
-          <button
-            className="secondary-action danger"
-            disabled={job.cancelRequested}
-            onClick={() => onCancel(job)}
-            type="button"
-          >
-            {job.cancelRequested ? "Canceling…" : "Cancel run"}
-          </button>
-          {onOpenQueue ? (
-            <button className="secondary-action" onClick={onOpenQueue} type="button">
-              View in Queue
-            </button>
-          ) : null}
-        </div>
-      ) : onOpenQueue ? (
-        <button className="secondary-action" onClick={onOpenQueue} type="button">
-          {isError ? "Open Queue" : "View in Queue"}
-        </button>
-      ) : null}
-    </article>
-  );
-}
